# services/api-sdk/models/google-auth-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Name
 * Poject Description description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface GoogleAuthDto
 */
export interface GoogleAuthDto {

    /**
     * The id token from Google OAuth2
     *
     * @type {string}
     * @memberof GoogleAuthDto
     * @example eyJhbGci..
     */
    id_token: string;
}