# services/api-sdk/models/manual-login-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Name
 * Poject Description description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ManualLoginDto
 */
export interface ManualLoginDto {

    /**
     * The email of the user
     *
     * @type {string}
     * @memberof ManualLoginDto
     * @example user@example.com
     */
    email: string;

    /**
     * The password of the user
     *
     * @type {string}
     * @memberof ManualLoginDto
     * @example password
     */
    password: string;
}