# services/api-sdk/models/manual-signup-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Name
 * Poject Description description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface ManualSignupDto
 */
export interface ManualSignupDto {

    /**
     * The name of the user
     *
     * @type {string}
     * @memberof ManualSignupDto
     * @example John Doe
     */
    name: string;

    /**
     * The email of the user
     *
     * @type {string}
     * @memberof ManualSignupDto
     * @example user@example.com
     */
    email: string;

    /**
     * The password of the user
     *
     * @type {string}
     * @memberof ManualSignupDto
     * @example password
     */
    password: string;

    /**
     * Password confirmation
     *
     * @type {string}
     * @memberof ManualSignupDto
     * @example password
     */
    confirm_password: string;
}