# services/api-sdk/models/role-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Name
 * Poject Description description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface RoleDto
 */
export interface RoleDto {

    /**
     * The role to assign to the user
     *
     * @type {string}
     * @memberof RoleDto
     * @example admin
     */
    role: string;
}