# services/server/storage/api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1": {
      "get": {
        "operationId": "AppController_getHello_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/v1/auth/google/verify": {
      "post": {
        "operationId": "AuthController_googleAuthVerify_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/signup": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualSignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/role": {
      "post": {
        "operationId": "AuthController_setRole_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/admin/profile": {
      "get": {
        "operationId": "AuthController_getAdminProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/user/profile": {
      "get": {
        "operationId": "AuthController_getUserProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "Project Name",
    "description": "Poject Description description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Project Tag",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "GoogleAuthDto": {
        "type": "object",
        "properties": {
          "id_token": {
            "type": "string",
            "example": "eyJhbGci..",
            "description": "The id token from Google OAuth2"
          }
        },
        "required": [
          "id_token"
        ]
      },
      "ManualSignupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "The name of the user"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "password",
            "description": "The password of the user"
          },
          "confirm_password": {
            "type": "string",
            "example": "password",
            "description": "Password confirmation"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "confirm_password"
        ]
      },
      "ManualLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email of the user"
          },
          "password": {
            "type": "string",
            "example": "password",
            "description": "The password of the user"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "admin",
            "description": "The role to assign to the user"
          }
        },
        "required": [
          "role"
        ]
      }
    }
  }
}